SensorEvent refactoring
