+++ src/main/java/ru/sbt/mipt/oop/SensorEventType.java
@@ -1,0 +1,5 @@
+package ru.sbt.mipt.oop;
+
+public enum SensorEventType {
+    DOOR_EVENT, LIGHT_EVENT
+}
