+++ src/main/java/ru/sbt/mipt/oop/eventfactories/EventFactory.java
@@ -1,6 +1,6 @@
 package ru.sbt.mipt.oop.eventfactories;
 
-import ru.sbt.mipt.oop.devices.SensorEvent;
+import ru.sbt.mipt.oop.SensorEvent;
 
 public interface EventFactory {
     SensorEvent getNextSensorEvent();
