+++ src/main/java/ru/sbt/mipt/oop/eventfactories/RandomEventFactory.java
@@ -1,10 +1,9 @@
 package ru.sbt.mipt.oop.eventfactories;
 
-import ru.sbt.mipt.oop.devices.SensorEvent;
+import ru.sbt.mipt.oop.SensorEvent;
+import ru.sbt.mipt.oop.SensorEventType;
 import ru.sbt.mipt.oop.devices.door.DoorActionType;
-import ru.sbt.mipt.oop.devices.door.DoorEvent;
 import ru.sbt.mipt.oop.devices.light.LightActionType;
-import ru.sbt.mipt.oop.devices.light.LightEvent;
 
 public class RandomEventFactory implements EventFactory {
 
@@ -14,9 +13,9 @@ public class RandomEventFactory implements EventFactory {
         String objectId = "" + ((int) (10 * Math.random()));
         int actionType = (int) (Math.random() * 2);
         if ((int) (Math.random() * 2) == 0) {
-            event = new DoorEvent(objectId, DoorActionType.values()[actionType]);
+            event = new SensorEvent(SensorEventType.DOOR_EVENT, DoorActionType.values()[actionType], objectId);
         } else {
-            event = new LightEvent(objectId, LightActionType.values()[actionType]);
+            event = new SensorEvent(SensorEventType.LIGHT_EVENT, LightActionType.values()[actionType], objectId);
         }
         return event;
     }
