+++ src/main/java/ru/sbt/mipt/oop/eventprocessors/EventProcessor.java
@@ -1,6 +1,6 @@
 package ru.sbt.mipt.oop.eventprocessors;
 
-import ru.sbt.mipt.oop.devices.SensorEvent;
+import ru.sbt.mipt.oop.SensorEvent;
 
 public interface EventProcessor {
     void process(SensorEvent event);
