+++ src/main/java/ru/sbt/mipt/oop/eventprocessors/processors/SetDoorState.java
@@ -1,11 +1,11 @@
 package ru.sbt.mipt.oop.eventprocessors.processors;
 
 import ru.sbt.mipt.oop.Room;
+import ru.sbt.mipt.oop.SensorEvent;
+import ru.sbt.mipt.oop.SensorEventType;
 import ru.sbt.mipt.oop.SmartHome;
-import ru.sbt.mipt.oop.devices.SensorEvent;
 import ru.sbt.mipt.oop.devices.door.Door;
 import ru.sbt.mipt.oop.devices.door.DoorActionType;
-import ru.sbt.mipt.oop.devices.door.DoorEvent;
 import ru.sbt.mipt.oop.eventprocessors.EventProcessor;
 import ru.sbt.mipt.oop.utils.Logger;
 
@@ -18,7 +18,7 @@ public class SetDoorState implements EventProcessor {
 
     @Override
     public void process(SensorEvent event) {
-        if (!(event instanceof DoorEvent)) return;
+        if (event.getEventType() != SensorEventType.DOOR_EVENT) return;
 
         for (Room room : smartHome.getRooms()) {
             for (Door door: room.getDoors()) {
