+++ src/main/java/ru/sbt/mipt/oop/utils/Logger.java
@@ -1,6 +1,6 @@
 package ru.sbt.mipt.oop.utils;
 
-import ru.sbt.mipt.oop.devices.SensorEvent;
+import ru.sbt.mipt.oop.SensorEvent;
 
 public class Logger {
     public static void newEvent(SensorEvent event) {
