--- src/main/test/ru/sbt/mipt/oop/eventprocessors/processors/SetDoorStateTest.java
@@ -1,15 +1,0 @@
-package ru.sbt.mipt.oop.eventprocessors.processors;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-class SetDoorStateTest {
-
-    @BeforeEach
-    void setUp() {
-    }
-
-    @Test
-    void process() {
-    }
-}
